feat: add form submission handler to IdentityForm component
This commit adds a form submission handler to the IdentityForm component in order to handle the submission of identity data. The handleSubmit function is modified to make an asynchronous POST request to the server endpoint 'http://localhost:3000/api/identities' with the entered identity data. After the request is made, the input field is cleared.

Ref: Issue #123